[Web] Refactor: :rainbow: change the location of socketController and socketService #452
소켓 컨트롤러와 소켓 서비스 파일의 위치를 변경한다.
