+++ server/src/sockets/socket.ts
@@ -1,7 +1,7 @@
 import https from 'https';
 import socketIO, { Socket } from 'socket.io';
 
-import socketController from './socketController';
+import socketController from '../controllers/socket';
 import { UserType, SendChatType } from '../@types/dataType';
 
 const socketLoader = (server: https.Server) => {
